Moved the prettifier classes and code to aiida.common.utils Improved 'simple' prettifiers to put digits to subscript
